fixing logzio file path
